uint32/uint16 -> uint256

+++ script/Deploy.s.sol
@@ -22,7 +22,7 @@ abstract contract DeployConstants is ReflexConstants {
     /**
      * @dev Module version of built-in upgradeable installer module.
      */
-    uint32 internal constant _MODULE_VERSION_INSTALLER = 1;
+    uint256 internal constant _MODULE_VERSION_INSTALLER = 1;
 
     /**
      * @dev Module upgradeability setting of built-in upgradeable installer module.
@@ -38,8 +38,8 @@ contract DeployScript is Script, DeployConstants {
     // Constants
     // =========
 
-    uint32 internal constant _MODULE_ID_EXAMPLE = 2;
-    uint32 internal constant _MODULE_VERSION_EXAMPLE = 1;
+    uint256 internal constant _MODULE_ID_EXAMPLE = 2;
+    uint256 internal constant _MODULE_VERSION_EXAMPLE = 1;
     bool internal constant _MODULE_UPGRADEABLE_EXAMPLE = true;
 
     // =======
