+++ test/fixtures/ReflexFixture.sol
@@ -55,9 +55,9 @@ abstract contract ReflexFixture is ReflexConstants, DeployConstants, TestHarness
 
     function _testModuleConfiguration(
         IReflexModule module_,
-        uint32 moduleId_,
-        uint16 moduleType_,
-        uint32 moduleVersion_,
+        uint256 moduleId_,
+        uint256 moduleType_,
+        uint256 moduleVersion_,
         bool moduleUpgradeable_
     ) internal {
         IReflexModule.ModuleSettings memory moduleSettings = module_.moduleSettings();
